fix: zero compiler/linker warnings on x86/x64 (BOM, 64-bit MemSearch, UINT codepage) + polish (favicon, og:image, repo meta)

+++ lib/Rxx/Str.h
@@ -10,14 +10,16 @@ namespace Rut
 	{
 		std::locale& GetCVT_UTF8();
 
-		std::wstring StrToWStr(const std::string& msString, size_t uCodePage);
-		std::string  WStrToStr(const std::wstring& wsString, size_t uCodePage);
-		size_t       StrToWStr(const std::string& msString, std::wstring& wsString, size_t uCodePage);
-		size_t       WStrToStr(const std::wstring& wsString, std::string& msString, size_t uCodePage);
+		// CodePage is fundamentally a UINT; using the exact type avoids size_t→UINT
+		// truncation warnings on 64-bit builds.
+		std::wstring StrToWStr(const std::string& msString, unsigned int uCodePage);
+		std::string  WStrToStr(const std::wstring& wsString, unsigned int uCodePage);
+		size_t       StrToWStr(const std::string& msString, std::wstring& wsString, unsigned int uCodePage);
+		size_t       WStrToStr(const std::wstring& wsString, std::string& msString, unsigned int uCodePage);
 
-		std::wstring StrToWStr_CVT(const std::string& msString, size_t uCodePage);
-		std::string  WStrToStr_CVT(const std::wstring& wsString, size_t uCodePage);
-		void         StrToWStr_CVT(const std::string& msString, std::wstring& wsString, size_t uCodePage);
-		void         WStrToStr_CVT(const std::wstring& wsString, std::string& msString, size_t uCodePage);
+		std::wstring StrToWStr_CVT(const std::string& msString, unsigned int uCodePage);
+		std::string  WStrToStr_CVT(const std::wstring& wsString, unsigned int uCodePage);
+		void         StrToWStr_CVT(const std::string& msString, std::wstring& wsString, unsigned int uCodePage);
+		void         WStrToStr_CVT(const std::wstring& wsString, std::string& msString, unsigned int uCodePage);
 	}
 }
